+++ src/Program.cs
@@ -2,7 +2,8 @@ using Microsoft.AspNetCore.Builder;
 using Bicep.Local.Extension.Host.Extensions;
 using Microsoft.Extensions.DependencyInjection;
 using Bicep.Extension.Databricks;
-using Bicep.Extension.Databricks.Handlers;
+using Bicep.Extension.Databricks.Handlers.Compute;
+using Bicep.Extension.Databricks.Handlers.Workspace;
 
 var builder = WebApplication.CreateBuilder();
 
@@ -15,7 +16,8 @@ builder.Services
         typeAssembly: typeof(Program).Assembly,
         configurationType: typeof(Configuration))
     .WithResourceHandler<SecretHandler>()
-    .WithResourceHandler<DirectoryHandler>();
+    .WithResourceHandler<DirectoryHandler>()
+    .WithResourceHandler<ClusterHandler>();
 
 var app = builder.Build();